requirements.txt
# Core TTS and audio processing
TTS>=0.21.1  # Coqui TTS
torch>=2.0.0  # PyTorch (CPU by default, add cuda version if you have GPU)
numpy>=1.20.0
librosa>=0.10.0  # Audio processing
soundfile>=0.12.1  # Audio I/O
ffmpeg-python>=0.2.0  # Video/audio conversion

# Development and testing
pytest>=7.0
black>=23.0.0  # Code formatting
isort>=5.12.0  # Import sorting
